tests/wpt/web-platform-tests/workers/constructors/Worker/resolve-empty-string.html

<!--
postMessage('ok');
/*
-->
<!doctype html>
<meta charset=utf-8>
<title>resolve the empty string</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var t = async_test();
t.step(function() {
  var worker = new Worker('');
  worker.addEventListener('message', t.step_func_done(function(e) {
    assert_equals(e.data, 'ok');
  }), false);
});
</script>
<!--
*/
//-->
